+++ R/pin_manifest.R
@@ -1,3 +1,5 @@
+# Variation on pin_registry use for versioning
+
 pin_manifest_get <- function(path) {
   manifest <- list()
 
@@ -30,6 +32,7 @@ pin_manifest_create <- function(path, metadata, files) {
 
   entries[sapply(entries, is.null)] <- NULL
 
+  fs::dir_create(path)
   yaml::write_yaml(entries, file.path(path, "data.txt"))
 }