+++ R/pin_tools.R
@@ -40,17 +40,6 @@ pin_results_extract_column <- function(df, column) {
   df
 }
 
-pin_reset_cache <- function(board, name) {
-  # clean up name in case it's a full url
-  name <- gsub("^https?://", "", name)
-
-  index <- tryCatch(pin_registry_retrieve(name, board), error = function(e) NULL)
-  if (!is.null(index)) {
-    index$cache <- list()
-    pin_registry_update(name, board, params = index)
-  }
-}
-
 pin_entries_to_dataframe <- function(entries) {
   jsonlite::fromJSON(jsonlite::toJSON(entries, null = "null", auto_unbox = TRUE))
 }